abtracted Flow validation code and successfully deployed trigger with new TOML format
