+++ packages/app/src/cli/models/extensions/specifications/flow_trigger.ts
@@ -1,64 +1,20 @@
 import {BaseSchema} from '../schemas.js'
-
 import {createExtensionSpecification} from '../specification.js'
 import {zod} from '@shopify/cli-kit/node/schema'
-
-const SUPPORTED_COMMERCE_OBJECTS = [
-  'customer_reference',
-  'order_reference',
-  'product_reference',
-  'marketing_activity_reference',
-  'abandonment_reference',
-]
-
-interface ConfigField {
-  type: string
-  required: boolean
-  key?: string | undefined
-  name?: string | undefined
-  description?: string | undefined
-}
-
-const validateCommerceObject = (configField: ConfigField) => {
-  if (!SUPPORTED_COMMERCE_OBJECTS.includes(configField.type)) {
-    if (!configField.key) {
-      throw new zod.ZodError([
-        {
-          code: zod.ZodIssueCode.custom,
-          path: ['settings.fields.key'],
-          message: 'Key must be speicified for non-commerce object fields',
-        },
-      ])
-    }
-
-    if (!configField.name) {
-      throw new zod.ZodError([
-        {
-          code: zod.ZodIssueCode.custom,
-          path: ['settings.fields.name'],
-          message: 'Name must be speicified for non-commerce object fields',
-        },
-      ])
-    }
-  }
-}
+import {serializeFields, validateCommerceObject} from '../../../services/Flow/validation.js'
 
 const FlowTriggerExtensionSchema = BaseSchema.extend({
   name: zod.string(),
   description: zod.string().optional(),
-  type: zod.literal('flow_action'),
+  type: zod.literal('flow_trigger'),
   extensions: zod
     .array(
       zod.object({
-        runtime_url: zod.string(),
-        validation_url: zod.string().optional(),
-        config_page_url: zod.string().optional(),
-        config_page_preview_url: zod.string().optional(),
         schema: zod.string().optional(),
         return_type_ref: zod.string().optional(),
       }),
     )
-    .min(1),
+    .optional(),
   settings: zod.object({
     fields: zod
       .array(
@@ -67,10 +23,9 @@ const FlowTriggerExtensionSchema = BaseSchema.extend({
             key: zod.string().optional(),
             name: zod.string().optional(),
             description: zod.string().optional(),
-            required: zod.boolean(),
             type: zod.string(),
           })
-          .refine(validateCommerceObject),
+          .refine((field) => validateCommerceObject(field, 'flow_trigger')),
       )
       .optional(),
   }),
@@ -85,10 +40,12 @@ const flowTriggerSpecification = createExtensionSpecification({
   singleEntryPath: false,
   appModuleFeatures: (_) => [],
   deployConfig: async (config, _) => {
+    console.log({fields: serializeFields('flow_trigger', config.settings.fields)})
+
     return {
       title: config.name,
       description: config.description,
-      fields: config.settings.fields,
+      fields: serializeFields('flow_trigger', config.settings.fields),
     }
   },
 })
